pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "ml_ops_project"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Group 39", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc-gs>=3.0.2",
    "fastapi==0.115.6",
    "google-api-python-client>=2.188.0",
    "httpx>=0.28.1",
    "hydra-core>=1.3.2",
    "locust>=2.43.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "pathspec<1.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "torch==2.6.0",
    "torchvision>=0.21.0",
    "typer==0.15.1",
    "uvicorn[standard]==0.34.0",
    "wandb>=0.24.0",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120
lint.select = ["I"]

[tool.coverage.run]
omit = ["tests/*"]
